Radar Lock and Cell Reorder Fix
This commit fixes the locked radar screen on initial launch. The bug was caused by the viewDidAppear method being called twice and thus resetting a boolean variable to lock the screen. I also fixed the reordering of the cells in the table view. I left some large blocks of commented out code that are no longer needed so feel free to make sure it works and delete that if you want.